# docs/swagger.yaml
basePath: /Vehicle
definitions:
  swagger.TechnicalService:
    properties:
      endDate:
        type: string
      fkTechnicalServiceType:
        type: integer
      fkVehicle:
        type: integer
      startDate:
        type: string
    type: object
  swagger.TechnicalServiceStatus:
    properties:
      name:
        type: string
    type: object
...
  swagger.TechnicalServiceType:
    properties:
      name:
        type: string
    type: object
...
host: 192.168.40.104:60030
info:
  contact: {}
  version: "1"
paths:
  /technicalService/:
    get:
      description: Obtiene una lista de todos servicios tecnicos.
      parameters:
      - default: 1
        description: Número de página para la paginación
        in: query
        name: page
        type: integer
      - default: 10
        description: Tamaño de página para la paginación
        in: query
        name: pageSize
        type: integer
      - description: Placa del Vehiculo
        in: query
        name: name
        type: string
      - description: Nombre del tipo de servicio
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene todos los  Servicios Tecnicos
      tags:
      - TechnicalService
    post:
      consumes:
      - application/json
      description: Crea un nuevio  servicio tecnico con los datos proporcionados
      parameters:
      - description: Datos de la servicio a crear
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.TechnicalService'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea una nuevo  servicio tecnico con los datos proporcionados
      tags:
      - TechnicalService
  /technicalServiceStatus/:
    get:
      description: Obtiene una lista de todos Estados de servicios tecnicos.
      parameters:
      - default: 1
        description: Número de página para la paginación
        in: query
        name: page
        type: integer
      - default: 10
        description: Tamaño de página para la paginación
        in: query
        name: pageSize
        type: integer
      - description: Nombre del Estado
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene todos los Estados de Servicios Tecnicos
      tags:
      - TechnicalServiceStatus
    post:
      consumes:
      - application/json
      description: Crea un nuevio estado de servicio tecnico con los datos proporcionados
      parameters:
      - description: Datos de la estado a crear
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.TechnicalServiceStatus'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea una nuevo estado de servicio tecnico con los datos proporcionados
      tags:
      - TechnicalServiceStatus
  /technicalServiceStatus/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un Estado de Servicio Tecnico  de la base de datos basado
        en el ID proporcionado
      parameters:
      - description: ID del Estado de Servicio tecnico a eliminar
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mensaje de éxito al eliminar
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tipo de  Servicio Tecnico no encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Elimina un Tipo de Estado de Servicio Tecnico por ID
      tags:
      - TechnicalServiceStatus
    get:
      description: Obtiene un Estado por su ID específico
      parameters:
      - description: ID del Estado
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalles del técnico
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: ID inválido'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: tecnico no encontrado'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene Estado de servicio tecnico en  específicoestado
      tags:
      - TechnicalServiceStatus
    put:
      consumes:
      - application/json
      description: Actualiza un tecnico por ID con los datos proporcionados
      parameters:
      - description: ID del Estado a actualizar
        in: path
        name: id
        required: true
        type: integer
      - description: Datos de la Estado a actualizar
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.TechnicalServiceStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Actualiza un Estado por ID
      tags:
      - TechnicalServiceStatus
...
  /technicalServiceType/:
    get:
      description: Obtiene una lista de todos los Tipos de Servicios Tecnicos.
      parameters:
      - default: 1
        description: Número de página para la paginación
        in: query
        name: page
        type: integer
      - default: 10
        description: Tamaño de página para la paginación
        in: query
        name: pageSize
        type: integer
      - description: Nombre del Tipo de servicio Tecnico
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene todos los Tipos de Servicios Tecnicos
      tags:
      - TechnicalServiceType
    post:
      consumes:
      - application/json
      description: Crea un nuevio tipo de servicio tecnico con los datos proporcionados
      parameters:
      - description: Datos de la marca de vehículo a crear
        in: body
        name: vb
        required: true
        schema:
          $ref: '#/definitions/swagger.TechnicalServiceType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea una nuevo tipo de servicio tecnico con los datos proporcionados
      tags:
      - TechnicalServiceType
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"